Temporarily make vmx an admin
If approved I'll merge it when I get the time to do the actual move, to keep the "I'm an admin" time as little as possible.

I just though I do a PR now to also clarify whether that's the correct way of doing it. @galargh: is this the correct way?